refactor(LineServcie): change to use VO
- VO로 바뀐 설계를 반영하기 위한 리팩토링
