+++ src/main/java/nextstep/subway/line/domain/exceptions/InvalidSectionException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.line.domain.exceptions;
+
+public class InvalidSectionException extends RuntimeException {
+    public InvalidSectionException(final String message) {
+        super(message);
+    }
+}
