+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -3,9 +3,9 @@ package nextstep.subway.line.ui;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.application.exceptions.LineNotFoundException;
 import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.exceptions.InvalidSectionException;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.section.domain.exceptions.InvalidSectionException;
 import nextstep.subway.station.domain.exceptions.StationNotExistException;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.http.ResponseEntity;
