--- src/main/java/nextstep/subway/section/application/SectionService.java
@@ -1,20 +1,0 @@
-package nextstep.subway.section.application;
-
-import nextstep.subway.section.domain.Section;
-import nextstep.subway.section.domain.SectionRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-public class SectionService {
-    private SectionRepository sectionRepository;
-
-    public SectionService(SectionRepository sectionRepository) {
-        this.sectionRepository = sectionRepository;
-    }
-
-    public Section createNewSection(Section section) {
-        return sectionRepository.save(section);
-    }
-}
