--- src/main/java/nextstep/subway/section/domain/exceptions/InvalidSectionException.java
@@ -1,7 +1,0 @@
-package nextstep.subway.section.domain.exceptions;
-
-public class InvalidSectionException extends RuntimeException {
-    public InvalidSectionException(String message) {
-        super(message);
-    }
-}
